Fix actionable advice imports for Jest
